server: move queries rpc from admin to status server
This will enable us to use the iterateNodes helper to fan out the request
to all the nodes in the cluster.

Release note: None
