+++ pkg/server/serverpb/admin.proto
@@ -17,7 +17,6 @@ package cockroach.server.serverpb;
 option go_package = "serverpb";
 
 import "config/zone.proto";
-import "roachpb/app_stats.proto";
 import "sql/jobs/jobspb/jobs.proto";
 import "storage/engine/enginepb/mvcc.proto";
 import "storage/liveness.proto";
@@ -549,15 +548,6 @@ message DataDistributionResponse {
   map<int64, ZoneConfig> zone_configs = 2 [(gogoproto.nullable) = false];
 }
 
-message QueriesRequest {
-}
-
-message QueriesResponse {
-  repeated cockroach.sql.CollectedStatementStatistics queries = 1 [(gogoproto.nullable) = false];
-   // Timestamp of the last stats reset.
-   google.protobuf.Timestamp last_reset = 3 [(gogoproto.nullable) = false, (gogoproto.stdtime) = true];
-}
-
   // MetricMetadataRequest requests metadata for all metrics.
 message MetricMetadataRequest {
 }
@@ -744,12 +734,6 @@ service Admin {
   rpc DataDistribution(DataDistributionRequest) returns (DataDistributionResponse) {
     option (google.api.http) = {
       get: "/_admin/v1/data_distribution"
-    };
-  }
-
-  rpc Queries(QueriesRequest) returns (QueriesResponse) {
-    option (google.api.http) = {
-      get: "/_admin/v1/queries"
     };
   }
