+++ pkg/server/serverpb/status.proto
@@ -18,6 +18,7 @@ option go_package = "serverpb";
 
 import "build/info.proto";
 import "gossip/gossip.proto";
+import "roachpb/app_stats.proto";
 import "roachpb/data.proto";
 import "server/diagnosticspb/diagnostics.proto";
 import "server/status/status.proto";
@@ -614,6 +615,15 @@ message StoresResponse {
   repeated StoreDetails stores = 1 [ (gogoproto.nullable) = false ];
 }
 
+message QueriesRequest {
+}
+
+message QueriesResponse {
+  repeated cockroach.sql.CollectedStatementStatistics queries = 1 [(gogoproto.nullable) = false];
+   // Timestamp of the last stats reset.
+   google.protobuf.Timestamp last_reset = 3 [(gogoproto.nullable) = false, (gogoproto.stdtime) = true];
+}
+
 service Status {
   rpc Certificates(CertificatesRequest) returns (CertificatesResponse) {
     option (google.api.http) = {
@@ -747,6 +757,12 @@ service Status {
   rpc Stores(StoresRequest) returns (StoresResponse) {
     option (google.api.http) = {
       get : "/_status/stores/{node_id}"
+    };
+  }
+
+  rpc Queries(QueriesRequest) returns (QueriesResponse) {
+    option (google.api.http) = {
+      get: "/_status/queries"
     };
   }
 }
