+++ pkg/server/status.go
@@ -1577,6 +1577,18 @@ func (s *statusServer) Stores(
 	return resp, nil
 }
 
+func (s *statusServer) Queries(
+	ctx context.Context, req *serverpb.QueriesRequest,
+) (*serverpb.QueriesResponse, error) {
+	stmtStats := s.admin.server.pgServer.SQLServer.GetUnscrubbedStmtStats()
+	lastReset := s.admin.server.pgServer.SQLServer.GetStmtStatsLastReset()
+
+	return &serverpb.QueriesResponse{
+		Queries:   stmtStats,
+		LastReset: lastReset,
+	}, nil
+}
+
 // jsonWrapper provides a wrapper on any slice data type being
 // marshaled to JSON. This prevents a security vulnerability
 // where a phishing attack can trick a user's browser into
